move smartInterval to utils
