+++ src/example-types/legacy/smartIntervalHistogram.js
@@ -1,6 +1,6 @@
 let _ = require('lodash/fp')
 let statsResults = require('../statistical').result
-let calcSmartInterval = require('../../smartInterval').calcSmartInterval
+let calcSmartInterval = require('../../utils/smartInterval').calcSmartInterval
 
 module.exports = {
   validContext: node => node.field,
