+++ src/utils/smartInterval.js
